app/auth/dependencies.py
import os
from fastapi import Depends, HTTPException
from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
from jose import jwt, JWTError
from psycopg2.extras import RealDictCursor

from app.db import get_conn

SECRET_KEY = os.getenv("JWT_SECRET", "dev-secret-local")
ALGORITHM = "HS256"
security = HTTPBearer()

def get_current_user(creds: HTTPAuthorizationCredentials = Depends(security)):
    token = creds.credentials
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        user_id = payload.get("sub")
        if not user_id:
            raise HTTPException(status_code=401, detail="Token inválido")
    except JWTError:
        raise HTTPException(status_code=401, detail="Token inválido")

    conn = get_conn()
    cur = conn.cursor(cursor_factory=RealDictCursor)
    cur.execute("SELECT id, email, role, phone_number_id, is_active FROM users WHERE id=%s", (user_id,))
    user = cur.fetchone()
    cur.close()
    conn.close()

    if not user or not user.get("is_active"):
        raise HTTPException(status_code=401, detail="Usuário inativo ou inexistente")

    return user
