requirements.txt
fastapi
uvicorn[standard]
jinja2
httpx
python-dotenv
psycopg2-binary

python-jose[cryptography]
passlib
email-validator
